+++ Employeemanagement-main/springapp/src/main/java/com/examly/springapp/repository/DocumentRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.examly.springapp.model.Document;
 
 @Repository
-public interface DocumentRepo extends JpaRepository<Document,Integer> {
+public interface DocumentRepo extends JpaRepository<Document,Long> {
                  
     Document findByDocumentName(String documentName);
